scripts/requirements.txt
# Core LLM integrations
anthropic>=0.21.0
openai>=1.10.0
ollama==0.3.0

# Supabase + utilities
supabase>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=4.0.0

# OCR & PDF processing
pytesseract==0.3.10
pdf2image==1.16.3
pdfplumber==0.10.3
PyMuPDF==1.23.8
Pillow==10.2.0
numpy==1.24.3
opencv-python==4.9.0.80

# AI classification & ML helpers
transformers==4.36.0
torch>=2.1.0
pandas==2.1.4
scikit-learn==1.3.2
tqdm==4.66.1

# CLI helpers
click==8.1.7
rich==13.7.1
